requirements.txt
fastapi
uvicorn
pynacl
qq-botpy
pillow
jmcomic

